sound + health
temp sounds are now implemented along with a health and damage system. Things that could be better: Player death sound does not seem to play anymore could be due to multiple things. Health bar decreases by 2 instead of one everytime hit, not sure if that is due to length of time being hit or a second invisible bullet.